+++ ICD.Connect.Cameras/Proxies/Controls/CameraControlApi.cs
@@ -11,21 +11,20 @@
 		public const string PROPERTY_MUTE_STATE = "MuteState";
 
 		public const string METHOD_PAN_STOP = "PanStop";
-		public const string METHOD_PAN_TILT = "PanTilt";
 		public const string METHOD_PAN_LEFT = "PanLeft";
 		public const string METHOD_PAN_RIGHT = "PanRight";
 		public const string METHOD_TILT_STOP = "TiltStop";
 		public const string METHOD_TILT_UP = "TiltUp";
 		public const string METHOD_TILT_DOWN = "TiltDown";
 		public const string METHOD_ZOOM_STOP = "ZoomStop";
-		public const string METHOD_ZOOM = "Zoom";
 		public const string METHOD_ZOOM_IN = "ZoomIn";
 		public const string METHOD_ZOOM_OUT = "ZoomOut";
 		public const string METHOD_GET_PRESETS = "GetPresets";
 		public const string METHOD_ACTIVATE_PRESET = "ActivatePreset";
 		public const string METHOD_STORE_PRESET = "StorePreset";
 		public const string METHOD_SET_MUTE = "SetCameraMute";
-		public const string METHOD_SEND_HOME = "SendHome";
+		public const string METHOD_ACTIVATE_HOME = "ActivateHome";
+		public const string METHOD_STORE_HOME = "StoreHome";
 
 		public const string HELP_EVENT_PRESETS_UPDATED = "Raised when the collection of presets is modified";
 		public const string HELP_EVENT_FEATURES_UPDATED = "Raised when the supported features list is updated";
@@ -46,6 +45,7 @@
 		public const string HELP_METHOD_ACTIVATE_PRESET = "Tells the camera to change its position to the given preset.";
 		public const string HELP_METHOD_STORE_PRESET = "Stores the cameras current position in the given preset index.";
 		public const string HELP_METHOD_SET_MUTE = "Sets if the camera mute state should be active.";
-		public const string HELP_METHOD_SEND_HOME = "Resets camera to its predefined home position.";
+		public const string HELP_METHOD_ACTIVATE_HOME = "Resets camera to its predefined home position.";
+		public const string HELP_METHOD_STORE_HOME = "Stores the current position as the home position.";
 	}
 }