+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+ - Added StoreHome method to cameras
+
 ## [4.3.0] - 2020-02-20
 ### Added
  - Vaddio driver supports returning to home position